Moved warehouse editor classes into editor.warehouse
