+++ src/main/java/application/desktop/DesktopApplication.java
@@ -3,7 +3,7 @@ package application.desktop;
 import application.desktop.ui.FontAwesomeIcon;
 import application.desktop.ui.components.ApplicationToolbar;
 import application.desktop.ui.components.common.Panel;
-import application.desktop.ui.components.editor.WarehouseEditor;
+import application.desktop.ui.components.editor.warehouse.WarehouseEditor;
 import imgui.ImFontConfig;
 import imgui.ImFontGlyphRangesBuilder;
 import imgui.ImGui;
@@ -17,13 +17,6 @@ import imgui.type.ImInt;
 import org.lwjgl.BufferUtils;
 import utils.Pair;
 import warehouse.Warehouse;
-import warehouse.storage.StorageUnit;
-import warehouse.storage.containers.InMemoryStorageUnitContainer;
-import warehouse.storage.strategies.MultiTypeStorageUnitStrategy;
-import warehouse.storage.strategies.SingleTypeStorageStrategy;
-import warehouse.tiles.Rack;
-import warehouse.tiles.ReceiveDepot;
-import warehouse.tiles.ShipDepot;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
