+++ src/main/java/application/desktop/ui/components/editor/warehouse/WarehouseEditor.java
@@ -1,4 +1,4 @@
-package application.desktop.ui.components.editor;
+package application.desktop.ui.components.editor.warehouse;
 
 import application.desktop.ui.components.common.Panel;
 import warehouse.Warehouse;
